Refactor the state to be immutable
One the main concerns that we address by writing this game with FP is state mutability. There is a bug however as the state here is not immutable. This is due to objects being passed by reference rather than getting a fresh copy. To combat this use the ES6 spread operator to create a new object, newState/initialState, and refactor the state to be flat and accessible at the top level of the object. Then deconstruct the state in each function with the necessary state attributes.